리팩토링: Bus클래스로 분리
