+++ app/src/main/java/com/example/timetogo/MainActivity.java
@@ -51,21 +51,15 @@ public class MainActivity extends AppCompatActivity {
     static String strweek = null;
     String data;
 
-    public ArrayList<String> busRouteList; //노선Id들의 리스트
-    public ArrayList<String> stationList; //정류소Id들의 리스트
-    public ArrayList<String> stationNmList; //정류소 이름들의 리스트
-    public ArrayList<String> stationNoList; //정류소 번호들의 리스트
-    public ArrayList<String> seqList; //정류소 순번들의 리스트
     TextView result;
-    static public String text = "";
 
     String[] data_split;
     static String bus1;
     static String station1;
 
-    ArrayList<String> items;
-    ArrayAdapter<String> adapter;
-    ListView listView;
+    static ArrayList<String> items;
+    static ArrayAdapter<String> adapter;
+    static ListView listView;
 
     private static final int GPS_ENABLE_REQUEST_CODE = 2001;
     private static final int PERMISSIONS_REQUEST_CODE = 100;
@@ -73,6 +67,8 @@ public class MainActivity extends AppCompatActivity {
 
     static double curLat;
     static double curLng;
+
+    public Bus bus = new Bus();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -152,16 +148,22 @@ public class MainActivity extends AppCompatActivity {
                         }
                         if(nWeek==5){
                             strweek="목요일";
-                            alarmTime(23,2,1);
-                            alarmTime(23,3,2);
-                            alarmTime(23,4,3);
-                            alarmTime(23,5,4);
-                            alarmTime(23,6,5);
-                            alarmTime(23,7,6);
-                            alarmTime(23,8, 7);
+                            alarmTime(21,41,1);
+                            alarmTime(21,42,2);
+                            alarmTime(21,43,3);
+                            alarmTime(21,44,4);
+                            alarmTime(21,45,5);
+                            alarmTime(21,7,6);
+                            alarmTime(21,8, 7);
                         }
                         if(nWeek==6){
                             strweek="금요일";
+                            alarmTime(18,25,1);
+                            alarmTime(18,26,2);
+                            alarmTime(18,27,3);
+                            alarmTime(18,28,4);
+                            alarmTime(18,29,5);
+                            alarmTime(18,30, 6);
                         }
                         if(nWeek==7){
                             strweek="토요일";
@@ -221,14 +223,13 @@ public class MainActivity extends AppCompatActivity {
                     final Calendar cal;
 
                     cal = Calendar.getInstance();
+                    String send = "";
                     hour = cal.get(Calendar.HOUR_OF_DAY);
                     min = cal.get(Calendar.MINUTE);
                     nWeek = cal.get(Calendar.DAY_OF_WEEK);
-                    //socket = new Socket("ec2-13-209-36-232.ap-northeast-2.compute.amazonaws.com", 7777);
-                    //out = new PrintWriter(socket.getOutputStream(), true);
-                    //in = new BufferedReader(new InputStreamReader(
-                      //      socket.getInputStream()));
-                    out.print(nWeek+","+hour+","+min+"\n");
+                    send = nWeek + "," + hour + "," + min + "\n";
+
+                    out.print(send);
                     out.flush();
                 } catch (Exception e) {
                     e.printStackTrace();
@@ -259,7 +260,8 @@ public class MainActivity extends AppCompatActivity {
                                     station1 = data_split[0];
                                 }
 
-                                showBusList(bus1, station1);
+                                bus.showBusList(bus1, station1);
+                                //showBusList(bus1, station1);
                                 NotificationSomethings();
 
                             }
@@ -272,30 +274,6 @@ public class MainActivity extends AppCompatActivity {
             }
         };
         worker.start();
-    }
-
-    public void showBusList(String bus, String station) {
-        busRouteList = new ArrayList<String>();
-        stationList = new ArrayList<String>();
-        stationNmList = new ArrayList<String>();
-        stationNoList = new ArrayList<String>();
-        seqList = new ArrayList<String>();
-        text = "";
-
-        //result.append("\nbus no: " + bus);
-        //result.append("\nstation: " + station);
-
-        getBusRouteList(bus);
-        getStationsByRouteList(busRouteList.get(0));
-        for (int i = 0; i < stationNoList.size(); i++) {
-            if (stationNoList.get(i).equals(station)) {
-                getBusAPI(stationList.get(i), busRouteList.get(0), seqList.get(i));
-            }
-        }
-       // result.append("\n"+text);
-
-        items.add(text);
-        adapter.notifyDataSetChanged();
     }
 
     public void NotificationSomethings() {
@@ -313,7 +291,7 @@ public class MainActivity extends AppCompatActivity {
                 .setContentTitle("Time to go!!!").setWhen(System.currentTimeMillis())
 
                 //.setContentText("상태바 드래그시 보이는 서브타이틀").setVibrate(new long[]{40,300})
-                .setContentText(text).setVibrate(new long[]{40,300})
+                .setContentText(bus.text).setVibrate(new long[]{40,300})
 
                 //출처: https://androphil.tistory.com/368?category=423967 [소림사의 홍반장!]
                 // 더 많은 내용이라서 일부만 보여줘야 하는 경우 아래 주석을 제거하면 setContentText에 있는 문자열 대신 아래 문자열을 보여줌
@@ -346,180 +324,6 @@ public class MainActivity extends AppCompatActivity {
 
         assert notificationManager != null;
         notificationManager.notify(1234, builder.build()); // 고유숫자로 노티피케이션 동작시킴
-
-    }
-
-
-
-    public void getBusAPI(String stId, String busRouteId, String ord) { //getLowArrInfoByStIdList
-        TextView status1 = (TextView)findViewById(R.id.result); //파싱된 결과확인
-
-        boolean in_rtNm = false, in_arrmsg1 = false, in_arrmsg2 = false;
-
-        String rtNm = null, arrmsg1 = null, arrmsg2 = null;
-
-        try {
-            URL url = new URL("http://ws.bus.go.kr/api/rest/arrive/getLowArrInfoByRoute?serviceKey=4Sc5MPbhBeWyQLVQ1JM9AqGWarV75Qk9hG%2FLuWVNl7%2B4hyHn8nP0mGrsxqDJHcgRdrSYT7sjblUpHpqjDcXdbw%3D%3D&stId=" + stId + "&busRouteId=" + busRouteId + "&ord=" + ord);
-            XmlPullParserFactory parserCreator = XmlPullParserFactory.newInstance();
-            XmlPullParser parser = parserCreator.newPullParser();
-
-            parser.setInput(url.openStream(), null);
-
-            int parserEvent = parser.getEventType();
-            System.out.println("파싱 시작합니다");
-
-            while(parserEvent != XmlPullParser.END_DOCUMENT) {
-                switch(parserEvent) {
-                    case XmlPullParser.START_TAG:
-                        if(parser.getName().equals("rtNm")) {
-                            in_rtNm = true;
-                        }
-                        if(parser.getName().equals("arrmsg1")) {
-                            in_arrmsg1 = true;
-                        }
-                        if(parser.getName().equals("arrmsg2")) {
-                            in_arrmsg2 = true;
-                        }
-                        break;
-                    case XmlPullParser.TEXT:
-                        if(in_rtNm) {
-                            rtNm = parser.getText();
-                            in_rtNm = false;
-                        }
-                        if(in_arrmsg1) {
-                            arrmsg1 = parser.getText();
-                            in_arrmsg1 = false;
-                        }
-                        if(in_arrmsg2) {
-                            arrmsg2 = parser.getText();
-                            in_arrmsg2 = false;
-                        }
-                        break;
-                    case XmlPullParser.END_TAG:
-                        if(parser.getName().equals("itemList")) {
-                            //status1.setText(status1.getText() + "버스번호: " + rtNm + "첫번째 전: " + arrmsg1 + "두번째 전" + arrmsg2+"\n");
-                            text = "버스번호: " + rtNm + "첫번째 전: " + arrmsg1 + "두번째 전" + arrmsg2+"\n";
-                        }
-                        break;
-                }
-                parserEvent = parser.next();
-            }
-
-        } catch(Exception e) {
-            status1.setText("에러발생");
-        }
-    }
-
-    public void getBusRouteList(String busRouteNm) { //버스 번호를 입력하면 busRouteId를 리턴한다
-
-        boolean in_busRouteId = false;
-
-        String busRouteId = null;
-
-        try {
-            URL url = new URL("http://ws.bus.go.kr/api/rest/busRouteInfo/getBusRouteList?serviceKey=4Sc5MPbhBeWyQLVQ1JM9AqGWarV75Qk9hG%2FLuWVNl7%2B4hyHn8nP0mGrsxqDJHcgRdrSYT7sjblUpHpqjDcXdbw%3D%3D&strSrch="+busRouteNm);
-            XmlPullParserFactory parserCreator = XmlPullParserFactory.newInstance();
-            XmlPullParser parser = parserCreator.newPullParser();
-
-            parser.setInput(url.openStream(), null);
-
-            int parserEvent = parser.getEventType();
-            System.out.println("파싱 시작합니다");
-
-            while(parserEvent != XmlPullParser.END_DOCUMENT) {
-                switch(parserEvent) {
-                    case XmlPullParser.START_TAG:
-                        if(parser.getName().equals("busRouteId")) {
-                            in_busRouteId = true;
-                        }
-                        break;
-                    case XmlPullParser.TEXT:
-                        if(in_busRouteId) {
-                            busRouteId = parser.getText();
-
-                            in_busRouteId = false;
-                        }
-                        break;
-                    case XmlPullParser.END_TAG:
-                        if(parser.getName().equals("itemList")) {
-                            busRouteList.add(busRouteId);
-                            break;
-                        }
-                        break;
-                }
-                parserEvent = parser.next();
-            }
-
-        } catch(Exception e) {
-            //error
-        }
-    }
-
-    public void getStationsByRouteList(String busRouteId) { //busRouteId를 넣으면 stationId를 리턴한다.
-
-        boolean in_station = false, in_stationNm = false, in_stationNo = false, in_seq = false;
-
-        String station = null, stationNm = null, stationNo = null, seq = null;
-
-        try {
-            URL url = new URL("http://ws.bus.go.kr/api/rest/busRouteInfo/getStaionByRoute?serviceKey=4Sc5MPbhBeWyQLVQ1JM9AqGWarV75Qk9hG%2FLuWVNl7%2B4hyHn8nP0mGrsxqDJHcgRdrSYT7sjblUpHpqjDcXdbw%3D%3D&busRouteId="+busRouteId);
-            XmlPullParserFactory parserCreator = XmlPullParserFactory.newInstance();
-            XmlPullParser parser = parserCreator.newPullParser();
-
-            parser.setInput(url.openStream(), null);
-
-            int parserEvent = parser.getEventType();
-            System.out.println("파싱 시작합니다");
-
-            while(parserEvent != XmlPullParser.END_DOCUMENT) {
-                switch(parserEvent) {
-                    case XmlPullParser.START_TAG:
-                        if(parser.getName().equals("station")) {
-                            in_station = true;
-                        }
-                        if(parser.getName().equals("stationNm")) {
-                            in_stationNm = true;
-                        }
-                        if(parser.getName().equals("stationNo")) {
-                            in_stationNo = true;
-                        }
-                        if(parser.getName().equals("seq")) {
-                            in_seq = true;
-                        }
-                        break;
-                    case XmlPullParser.TEXT:
-                        if(in_station) {
-                            station = parser.getText();
-                            in_station = false;
-                        }
-                        if(in_stationNm) {
-                            stationNm = parser.getText();
-                            in_stationNm = false;
-                        }
-                        if(in_stationNo) {
-                            stationNo = parser.getText();
-                            in_stationNo = false;
-                        }
-                        if(in_seq) {
-                            seq = parser.getText();
-                            in_seq = false;
-                        }
-                        break;
-                    case XmlPullParser.END_TAG:
-                        if(parser.getName().equals("itemList")) {
-                            stationList.add(station);
-                            stationNmList.add(stationNm);
-                            stationNoList.add(stationNo);
-                            seqList.add(seq);
-                        }
-                        break;
-                }
-                parserEvent = parser.next();
-            }
-
-        } catch(Exception e) {
-            //error
-        }
     }
 
     @Override
